add distibuted

+++ src/tracker/async_local_tracker.h
@@ -60,7 +60,7 @@ class AsyncLocalTracker {
   void Wait(int num_remains = 0) {
     std::unique_lock<std::mutex> lk(mu_);
     fin_cond_.wait(lk, [this, num_remains] {
-        return pending_.size() + running_.size() <= num_remains;
+        return pending_.size() + running_.size() <= size_t(num_remains);
       });
   }
   /**
